# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Bookled Documentation",
        "description": "Bookled Documentation has everything you need to understand each endpoint",
        "version": "1.0.0",
        "contact": {
            "name": "Emad Mahmoud",
            "url": "https://www.linkedin.com/in/emadmahmoud/",
            "email": "emad.mahmoud.dev@gmail.com"
        }
    },
...
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local server"
        },
        {
            "url": "https://bookled.herokuapp.com",
            "description": "Production server"
        }
    ],
...
    "paths": {
...
        "/api/v1/category/add": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Add new category",
                "description": "Add new category to the database",
                "operationId": "addCategory",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Category added successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Category Created Successfully",
                                    "category": {
                                        "id": 1,
                                        "name": "Science",
                                        "description": "A science book is a work of nonfiction, usually written by a scientist, researcher, or professor.",
                                        "books": [],
                                        "user_id": "1",
                                        "_id": "657e720637fc1d3b85ba07fe",
                                        "__v": 0
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Validatoin Failed"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Not Authorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Internal server error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/grantBook/add": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add new grant book to User",
                "description": "Add new grant book to a specific user",
                "operationId": "addGrantBook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GrantBook"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Grant book added successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Grant Book Created Successfully",
                                    "grantBook": {
                                        "book_id": "65b295badd6d5358908e14f7",
                                        "grantCode": "47283405423"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Validatoin Failed"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Not Authorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found or User not found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Book not found"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Internal server error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "required": [
                    "name",
                    "description"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Category name",
                        "example": "Science"
                    },
                    "description": {
                        "type": "string",
                        "description": "Category description",
                        "example": "A science book is a work of nonfiction, usually written by a scientist, researcher, or professor."
                    }
                }
            },
            "GrantBook": {
                "required": [
                    "bookId",
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "bookId": {
                        "type": "string",
                        "description": "Book id",
                        "example": "65b295badd6d5358908e14f7"
                    },
                    "grantCode": {
                        "type": "string",
                        "description": "Grant code",
                        "example": "123456"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}